Control de stock en los pedidos.

+++ application/controllers/ctrl_carrito.php
@@ -56,6 +56,7 @@ class Ctrl_carrito extends CI_Controller {
 	//Finaliza el pedido e inserta en la tabla de datos
 	public function FinalizaPedido(){
 		$carrito = new Carrito();
+		$interrumpir = false;
 		if($carrito->articulos_total()==0){//No hay productos en el carrito, lleva al index
 			redirect(base_url());
 		}
@@ -68,39 +69,59 @@ class Ctrl_carrito extends CI_Controller {
 			$this->load->model('model_productos');
 			//Obtiene los datos del usuario para introducirlos en el pedido
 			$usuario = $this->model_user->UsuarioID($this->session->userdata('id'));
-
-			$datos = array(
-				'id_usuario' => $usuario['id'],
-				'fecha' => date('Y-m-d h:i:s a', time()),
-				'nombre' => $usuario['nombre'],
-				'apellidos' => $usuario['apellidos'],
-				'direccion' => $usuario['direccion'],
-				'cp' => $usuario['cp'], 
-				'dni' => $usuario['dni'], 
-				'correo' => $usuario['correo'], 
-				'estado' => 'p' 
-
-				);
-			//Inserta el pedido y obtiene id de pedido
-			$pedidoid = $this->model_pedidos->InsertaPedido($datos);
-
-			//Se insertan todos los productos del carrito
 			$carro = $carrito->get_content();
 			foreach($carro as $producto){
-				$datos_linea = array(
-					'id_pedido' => $pedidoid,
-					'id_producto' => $producto['id'],
-					'precio' => $producto['precio'],
-					'cantidad' => $producto['cantidad']  
-					);
-				$this->model_pedidos->InsertaLinea($datos_linea);
-				$this->model_productos->BajaStock($producto['id'], $producto['cantidad']);
+				$stock = $this->model_productos->Stock($producto['id']);
+				if($producto['cantidad'] > $stock){
+					$mensaje = '<i class="fa fa-exclamation-triangle" aria-hidden="true"></i> No hay suficiente stock de '.$producto['nombre'].', por favor, introduzca una cantidad menor.';
+					$interrumpir = true; //Se interrumpe la operación de compra, ya que no hay stock suficiente
+					break;
+				}
 			}
-			//Una vez ha terminado el pedido, se vacía el carrito
-			$carrito->destroy();
 			
-			//Vista de éxito
-			$this->load->CargaVista('carrito/v_realizado',null);
+			if($interrumpir == false){
+					$datos = array(
+					'id_usuario' => $usuario['id'],
+					'fecha' => date('Y-m-d h:i:s a', time()),
+					'nombre' => $usuario['nombre'],
+					'apellidos' => $usuario['apellidos'],
+					'direccion' => $usuario['direccion'],
+					'cp' => $usuario['cp'], 
+					'dni' => $usuario['dni'], 
+					'correo' => $usuario['correo'], 
+					'estado' => 'p' 
+
+					);
+				//Inserta el pedido y obtiene id de pedido
+				$pedidoid = $this->model_pedidos->InsertaPedido($datos);
+
+				//Se insertan todos los productos del carrito
+				
+				foreach($carro as $producto){
+					
+
+					$datos_linea = array(
+						'id_pedido' => $pedidoid,
+						'id_producto' => $producto['id'],
+						'precio' => $producto['precio'],
+						'cantidad' => $producto['cantidad']  
+						);
+					//Se inserta la linea en la base de datos
+					$this->model_pedidos->InsertaLinea($datos_linea);
+					//Baja el stock del producto
+					$this->model_productos->BajaStock($producto['id'], $producto['cantidad']);
+				}
+
+				//Una vez ha terminado el pedido, se vacía el carrito
+				$carrito->destroy();
+			
+				//Vista de éxito
+				$this->load->CargaVista('carrito/v_realizado',null);
+			}
+			else{
+				$this->load->CargaVista('carrito/v_carrito', array('mensaje' =>  $mensaje , 'carrito'=>$carrito));
+			}
+			
 		}
 	}
 }
